Major change to Plasma desktop theme.
Replaced some coloured icons with a monochrome version. The purpose
of this is to have monochrome icons on all icons located on panels
and menus. Tooltip and notification icons that display a larger
version of the icon will remain coloured. This is to create a more
consistent feel of the Plasma/Kvantum theme and the Papirus icon theme.